==> 03_EDA_and_TSA.ipynb <==
%%capture
%run 02_FeatEng.ipynb

import mle.tsa as mle_tsa

ax = data.plot(y=['cons_GNA95', 'price_GNA95'], secondary_y='price_GNA95')
ax.set_ylabel('(t)')
ax.right_ax.set_ylabel("(€/l)")
plt.title('GNA95: Consumo y precio medio mensual')
plt.show()

# Seasonality Analysis

ax = mle_tsa.plot_time_series_maves(x=data['cons_GNA95'], periods=[3,12])

ax = mle_tsa.plot_time_series_maves(x=data['cons_GNA98_diff1_sdiff12'], periods=[3,12])

ax = mle_tsa.plot_time_series_with_rolling(x=data['cons_GNA95'], period=3)

fig, axs = mle_tsa.plot_seas_model(seas_models['cons_GNA95'], title='GNA95 consumption', ylabel='(tons)')
plt.show()

fig, axs = mle_tsa.plot_seas_model(seas_models['price_GNA95'], title='GNA95 price', ylabel='(€/l)')
plt.show()

# Stationarity Checking

adf_test_cols = ['cons_GNA95', 'cons_GNA95_pct_change1', 'cons_GNA95_seas_adj', 'cons_GNA95_seas_adj_pct_change1', 'cons_GNA95_seas_adj_diff1',
                 'cons_GNA95_diff1', 'cons_GNA95_diff1_sdiff12',
                'price_GNA95', 'price_GNA95_pct_change1', 'price_GNA95_seas_adj', 'price_GNA95_seas_adj_pct_change1', 'price_GNA95_seas_adj_diff1',
                'price_GNA95_diff1', 'price_GNA95_diff1_sdiff12',]
_ = mle_tsa.wrap_adfuller(data[adf_test_cols], regression='c')

_ = mle_tsa.wrap_adfuller(data[['cons_GNA95_seas_adj_pct_change1']], regression='ct')

_ = mle_tsa.wrap_kpss(data[['cons_GNA95_seas_adj_pct_change1']], regression='c', nlags='auto')

ax = mle_tsa.plot_time_series_with_rolling(x=data['price_GNA95_diff1_sdiff12'], period=12)

ax = mle_tsa.plot_time_series_with_rolling(x=data['price_GNA95_seas_adj_pct_change1'], period=12)

ax = mle_tsa.plot_time_series_with_rolling(x=data['cons_GNA95_seas_adj_pct_change1'], period=12)

ax = mle_tsa.plot_time_series_with_rolling(x=data['cons_GNA95_diff1_sdiff12'], period=12)

# Autocorrelation

fig, axs   = plt.subplots(2, 1, figsize=(16,12))

plot_acf(data['cons_GNA95_diff1'], lags=24, ax=axs[0])
plot_acf(data['cons_GNA95_diff1_sdiff12'],lags=24, ax=axs[1])
plt.show()

fig, axs   = plt.subplots(2, 1, figsize=(16,12))

plot_acf(data['price_GNA95_diff1'], lags=24, ax=axs[0])
plot_acf(data['price_GNA95_diff1_sdiff12'],lags=24, ax=axs[1])
plt.show()

fig, axs   = plt.subplots(2, 2, figsize=(16,12))

for idx_row, col in enumerate(['price_GNA95_seas_adj_pct_change1', 'cons_GNA95_seas_adj_pct_change1']):
    plot_acf(data[col], lags=24, ax=axs[idx_row,0])
    plot_pacf(data[col],lags=24, ax=axs[idx_row,1])
    axs[idx_row,0].set_ylabel(col)

gna_corr = mle_tsa.pairwise_corr_over_time(data, 'price_GNA95_seas_adj_pct_change1', ['cons_GNA95_seas_adj_pct_change1'], 3)
gna_corr.plot.barh()
plt.show()

g = sns.jointplot('price_GNA95_seas_adj_pct_change1_lag1', 'cons_GNA95_seas_adj_pct_change1', data, kind='reg')
plt.show()

week_gna = data.groupby('week')[['price_GNA95', 'cons_GNA95']].median()
g = sns.jointplot('price_GNA95', 'cons_GNA95', week_gna, kind='reg')
plt.show()
